Close the crossing window where it actually is, not around the call
Corrects 432303ca, which fixed the reset and introduced a freeze.

That commit put the cli in the guest's C wrapper, around the call into the
switch. The call does not merely cross: the host runs an entire monitor-loop
iteration inside it and re-enters the guest before it returns. So the cli was
held across all of that, and any path that leaves the switch without coming
back through the matching popfq -- a resume after a fault, an interrupt
replayed and the guest restarted at its saved frame -- put the guest back on
the processor with real interrupts off. The guest then ran free and deaf until
something happened to call local_irq_enable(). Delivering a signal and
respawning a shell is long enough for the host to miss its clock, so ctrl-C in
the guest's terminal froze the machine.

The window that needs covering is five instructions wide and is inside the
switch: between the CR3 write and the lidt, the processor is already in the
entering side's address space while IDTR still points at the leaving side's
table. It is closed with a bare cli, and nothing is saved or restored, because
at that point there is nowhere to put it -- RSP still belongs to the leaving
side until the stack switch further down, so a push touches memory the entering
space need not map, and r11, the obvious scratch, is claimed by the ltr
sequence below. Both were tried and rejected by reading rather than by another
freeze.

Nothing is lost by not restoring, because each side owns its own interrupt
state on the far side: the monitor loop puts the host's saved flags back when
the switch returns to it, and the guest re-enables explicitly after its yield.

That guest-side sti is the other half, and it fixes something that was already
true and merely lucky. A guest has always resumed from a voluntary yield with
real interrupts off, because the monitor loop clears IF before entering; what
covered it was Linux calling local_irq_enable() soon afterwards. A guest must
always run with real interrupts on, so it now says so rather than depending on
when the next enable happens to arrive.

Also here, from the same session:

The 4 GB refusal at driver load is gone, with the accounting behind it widened
to 64-bit -- hostmem_* were "unsigned long", which is 32 bits on Windows x64,
so hostmem_usage_limit <<= 20 would have silently wrapped to about 3.9 GB on a
bigger machine. Nothing in the x86-64 path needed the limit: physical addresses
are already 64-bit, guest RAM is allocated with no ceiling, and guest physical
equals host physical.

The reversed physical mapping is no longer built. It serves the i386 monitor
path, which this port does not use, and cost one page per 512 pages of host RAM
at load -- 8 MB per 4 GB, for a table nothing here reads.

And the terminal learned four keys: backspace, which Windows sends as BS while
a Unix line discipline erases on DEL, so the byte was being typed rather than
obeyed; plus Delete, Home and End, which went through the same extended-key
path as the arrows and were being dropped.

Verified on the M92p with both triggers that produced the two failures: a USB
KVM switched away and back with the guest idling at a prompt, and ctrl-C
pressed repeatedly in the terminal.

+++ src/colinux/kernel/manager.h
@@ -53,10 +53,17 @@ typedef struct co_manager_open_desc {
 typedef struct co_manager {
 	co_manager_state_t state;
 
-	unsigned long hostmem_amount;
-	unsigned long hostmem_used;
-	unsigned long hostmem_usage_limit;
-	unsigned long hostmem_pages;
+	/*
+	 * 64-bit because "unsigned long" is 32 bits on Windows x64, and these
+	 * are byte counts and page counts for the whole machine. The old
+	 * "more than 4GB is not supported" check made that safe by refusing to
+	 * load at all; with the check gone, hostmem_usage_limit <<= 20 on an
+	 * 8 GB box would overflow and silently come back as about 3.9 GB.
+	 */
+	unsigned long long hostmem_amount;
+	unsigned long long hostmem_used;
+	unsigned long long hostmem_usage_limit;
+	unsigned long long hostmem_pages;
 
 	co_pfn_t *reversed_map_pfns;
 	unsigned long reversed_page_count;
